refactor(ui): relocate wideMode helper to @plannotator/ui/utils (Phase 7 step 3)
